ExceptionHandling added.

+++ src/main/java/com/api/services/ProductService.java
@@ -1,12 +1,14 @@
 package com.api.services;
 
 import com.api.entity.Product;
+import com.api.exceptions.CategoryNotFoundException;
+import com.api.exceptions.ProductNotFoundException;
 import com.api.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -17,6 +19,9 @@ public class ProductService {
     public ProductService(ProductRepository productRepository) {
         this.productRepository = productRepository;
     }
+
+    @Autowired
+    CategoryService categoryService;
 
     public Product saveProduct(Product product){
         return productRepository.save(product);
@@ -30,31 +35,40 @@ public class ProductService {
         return productRepository.findAll();
     }
 
-    public Product getProductById(long id){
-        return productRepository.findById(id).orElseThrow(ResourceNotFoundException::new);
+    public Product getProductById(long id) throws ProductNotFoundException {
+        Optional<Product> product = productRepository.findById(id);
+        if (!product.isPresent())
+            throw new ProductNotFoundException("Product Not Found in Product Repository ProductId:" + id);
+
+        return product.get();
     }
 
     public Product getProductByName(String name){
         return productRepository.findByName(name);
     }
 
-    public Product updateProduct(Product product) {
-        Product existingProduct = getProductById(product.getId());
+    public Product updateProduct(long id, Product product) throws ProductNotFoundException, CategoryNotFoundException {
+        Optional<Product> optionalProduct = productRepository.findById(id);
+        if (!optionalProduct.isPresent())
+            throw new ProductNotFoundException("Product Not Found in Product Repository ProductId:" + id + ", Provide The Correct ProductId!");
 
-        if (existingProduct != null){
-            existingProduct.setUpdatedBy(product.getUpdatedBy());
-            existingProduct.setName(product.getName());
-            existingProduct.setQuantity(product.getQuantity());
-            existingProduct.setPrice(product.getPrice());
-            existingProduct.setCategory(product.getCategory());
-        }
+        categoryService.getCategoryById(product.getCategory().getId());
+        optionalProduct.get().setName(product.getName());
+        optionalProduct.get().setPrice(product.getPrice());
+        optionalProduct.get().setQuantity(product.getQuantity());
+        optionalProduct.get().setUpdatedBy(product.getUpdatedBy());
+        optionalProduct.get().setCategory(product.getCategory());
 
-        assert existingProduct != null;
-        return productRepository.save(existingProduct);
+        return productRepository.save(optionalProduct.get());
     }
 
-    public String deleteProduct(long id){
+    public String deleteProduct(long id) throws ProductNotFoundException {
+        Optional<Product> optionalProduct = productRepository.findById(id);
+        if (!optionalProduct.isPresent())
+            throw new ProductNotFoundException("Product Not Found in Product Repository ProductId:" + id + ", Provide The Correct ProductId!");
+
         productRepository.deleteById(id);
-        return "Product Removed ! Product ID:" + id;
+
+        return "Product Removed! Product ID:" + id;
     }
 }
